phase2/requirements.txt
psycopg2==2.9.10
pymongo==4.10.1
python-dotenv==1.0.1
